move provider definitions out of init
